+++ README.md
@@ -1,1 +1,10 @@
 # Mist
+
+
+## setup dev envirvonment
+
+```bash
+git clone https://github.com/corecollectives/mist
+cd mist
+docker-compose up --build
+```
