--- server/api/routes.go
@@ -1,10 +1,0 @@
-package api
-
-import (
-	"github.com/corecollectives/mist/api/handlers"
-	"net/http"
-)
-
-func RegisterRoutes(mux *http.ServeMux) {
-	mux.HandleFunc("/health", handlers.HealthCheckHandler)
-}
