+++ server/go.mod
@@ -1,3 +1,5 @@
 module github.com/corecollectives/mist
 
 go 1.25.1
+
+require github.com/mattn/go-sqlite3 v1.14.32 // indirect
